Separate web-services for different features by packages
