+++ app/src/main/java/com/github/fo2rist/mclaren/dagger/FeedRepositoriesModule.kt
@@ -6,12 +6,12 @@ import com.github.fo2rist.mclaren.repository.feed.McLarenFeedRepositoryImpl
 import com.github.fo2rist.mclaren.repository.RepositoryEventBusImpl
 import com.github.fo2rist.mclaren.repository.feed.StoryStreamRepositoryImpl
 import com.github.fo2rist.mclaren.repository.feed.TwitterRepositoryImpl
-import com.github.fo2rist.mclaren.web.FeedHistoryPredictor
-import com.github.fo2rist.mclaren.web.McLarenFeedHistoryPredictor
-import com.github.fo2rist.mclaren.web.McLarenFeedWebService
-import com.github.fo2rist.mclaren.web.McLarenWebServiceImpl
-import com.github.fo2rist.mclaren.web.StoryStreamWebService
-import com.github.fo2rist.mclaren.web.StoryStreamWebServiceImpl
+import com.github.fo2rist.mclaren.web.feed.FeedHistoryPredictor
+import com.github.fo2rist.mclaren.web.feed.McLarenFeedHistoryPredictor
+import com.github.fo2rist.mclaren.web.feed.McLarenFeedWebService
+import com.github.fo2rist.mclaren.web.feed.McLarenWebServiceImpl
+import com.github.fo2rist.mclaren.web.feed.StoryStreamWebService
+import com.github.fo2rist.mclaren.web.feed.StoryStreamWebServiceImpl
 import dagger.Binds
 import dagger.Module
 import javax.inject.Named
