+++ app/src/main/java/com/github/fo2rist/mclaren/dagger/RemoteConfigRepositoryModule.kt
@@ -4,8 +4,8 @@ import com.github.fo2rist.mclaren.repository.remoteconfig.DriversRepository
 import com.github.fo2rist.mclaren.repository.remoteconfig.DriversRepositoryImpl
 import com.github.fo2rist.mclaren.repository.remoteconfig.RaceCalendarRepository
 import com.github.fo2rist.mclaren.repository.remoteconfig.RaceCalendarRepositoryImpl
-import com.github.fo2rist.mclaren.web.FirebaseRemoteConfigService
-import com.github.fo2rist.mclaren.web.RemoteConfigService
+import com.github.fo2rist.mclaren.web.remoteconfig.FirebaseRemoteConfigService
+import com.github.fo2rist.mclaren.web.remoteconfig.RemoteConfigService
 import dagger.Binds
 import dagger.Module
 import javax.inject.Singleton
