+++ app/src/main/java/com/github/fo2rist/mclaren/dagger/TransmissionRepositoryModule.kt
@@ -4,8 +4,8 @@ import com.github.fo2rist.mclaren.repository.RepositoryEventBusImpl
 import com.github.fo2rist.mclaren.repository.transmission.TransmissionRepository
 import com.github.fo2rist.mclaren.repository.transmission.TransmissionRepositoryEventBus
 import com.github.fo2rist.mclaren.repository.transmission.TransmissionRepositoryImpl
-import com.github.fo2rist.mclaren.web.McLarenWebServiceImpl
-import com.github.fo2rist.mclaren.web.TransmissionWebService
+import com.github.fo2rist.mclaren.web.feed.McLarenWebServiceImpl
+import com.github.fo2rist.mclaren.web.transmission.TransmissionWebService
 import dagger.Binds
 import dagger.Module
 import javax.inject.Singleton
