+++ app/src/main/java/com/github/fo2rist/mclaren/ui/feedscreen/FeedAdapter.java
@@ -22,8 +22,8 @@ import com.github.fo2rist.mclaren.models.ImageUrl;
 import com.github.fo2rist.mclaren.models.Size;
 import com.github.fo2rist.mclaren.mvp.FeedContract.OnFeedInteractionListener;
 import com.github.fo2rist.mclaren.mvp.FeedContract.OnFeedScrollingListener;
-import com.github.fo2rist.mclaren.web.McLarenImageDownloader;
-import com.github.fo2rist.mclaren.web.McLarenImageDownloader.ImageSizeLimit;
+import com.github.fo2rist.mclaren.ui.tools.McLarenImageDownloader;
+import com.github.fo2rist.mclaren.ui.tools.McLarenImageDownloader.ImageSizeLimit;
 import com.luseen.autolinklibrary.AutoLinkMode;
 import com.luseen.autolinklibrary.AutoLinkOnClickListener;
 import com.luseen.autolinklibrary.AutoLinkTextView;
