+++ app/src/main/java/com/github/fo2rist/mclaren/web/feed/BaseFeedWebService.kt
@@ -1,4 +1,4 @@
-package com.github.fo2rist.mclaren.web
+package com.github.fo2rist.mclaren.web.feed
 
 import android.support.annotation.VisibleForTesting
 import com.github.fo2rist.mclaren.web.utils.executeSuspend
