+++ app/src/main/java/com/github/fo2rist/mclaren/web/feed/FeedHistoryPredictor.java
@@ -1,4 +1,4 @@
-package com.github.fo2rist.mclaren.web;
+package com.github.fo2rist.mclaren.web.feed;
 
 public interface FeedHistoryPredictor {
     int getFirstHistoryPage();
