+++ app/src/main/java/com/github/fo2rist/mclaren/web/remoteconfig/RemoteConfigService.kt
@@ -1,4 +1,4 @@
-package com.github.fo2rist.mclaren.web
+package com.github.fo2rist.mclaren.web.remoteconfig
 
 import com.github.fo2rist.mclaren.BuildConfig
 import com.github.fo2rist.mclaren.R
