+++ app/src/test/java/com/github/fo2rist/mclaren/repository/feed/McLarenFeedRepositoryImplTest.kt
@@ -1,11 +1,9 @@
 package com.github.fo2rist.mclaren.repository.feed
 
-import com.github.fo2rist.mclaren.repository.feed.FeedRepositoryEventBus
 import com.github.fo2rist.mclaren.repository.feed.FeedRepositoryEventBus.LoadingEvent
-import com.github.fo2rist.mclaren.repository.feed.McLarenFeedRepositoryImpl
 import com.github.fo2rist.mclaren.testdata.McLarenFeedResponse.REAL_FEED_RESPONSE
-import com.github.fo2rist.mclaren.web.FeedHistoryPredictor
-import com.github.fo2rist.mclaren.web.McLarenFeedWebService
+import com.github.fo2rist.mclaren.web.feed.FeedHistoryPredictor
+import com.github.fo2rist.mclaren.web.feed.McLarenFeedWebService
 import com.github.fo2rist.mclaren.web.utils.BadResponse
 import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.doReturn
