+++ app/src/test/java/com/github/fo2rist/mclaren/web/feed/McLarenFeedHistoryPredictorTest.kt
@@ -1,4 +1,4 @@
-package com.github.fo2rist.mclaren.web
+package com.github.fo2rist.mclaren.web.feed
 
 import com.github.fo2rist.mclaren.BuildConfig
 import com.github.fo2rist.mclaren.web.utils.BadResponse
