+++ app/src/test/java/com/github/fo2rist/mclaren/web/feed/StoryStreamWebServiceImplTest.kt
@@ -1,5 +1,8 @@
-package com.github.fo2rist.mclaren.web
+package com.github.fo2rist.mclaren.web.feed
 
+import com.github.fo2rist.mclaren.web.feed.BaseFeedWebService
+import com.github.fo2rist.mclaren.web.feed.BaseFeedWebServiceTest
+import com.github.fo2rist.mclaren.web.feed.StoryStreamWebServiceImpl
 import org.junit.Before
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
