+++ app/src/test/java/com/github/fo2rist/mclaren/web/models/StoryStreamParserTest.kt
@@ -1,7 +1,7 @@
 package com.github.fo2rist.mclaren.web.models
 
 import com.github.fo2rist.mclaren.testdata.StoryStreamResponse
-import com.github.fo2rist.mclaren.web.SafeJsonParser
+import com.github.fo2rist.mclaren.web.utils.SafeJsonParser
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
